move diffs out of scm provider (WIP)
Add file to filelog item
