+++ src/ScmProvider.ts
@@ -18,12 +18,7 @@ import { IPerforceConfig, matchConfig } from "./PerforceService";
 import * as Path from "path";
 import * as fs from "fs";
 import { WorkspaceConfigAccessor } from "./ConfigService";
-
-enum DiffType {
-    WORKSPACE_V_DEPOT,
-    SHELVE_V_DEPOT,
-    WORKSPACE_V_SHELVE
-}
+import * as DiffProvider from "./DiffProvider";
 
 export class PerforceSCMProvider {
     private wksFolder: Uri;
@@ -273,7 +268,10 @@ export class PerforceSCMProvider {
         const promises = [];
         for (const resource of selection) {
             promises.push(
-                PerforceSCMProvider.open(resource, DiffType.WORKSPACE_V_SHELVE)
+                PerforceSCMProvider.open(
+                    resource,
+                    DiffProvider.DiffType.WORKSPACE_V_SHELVE
+                )
             );
         }
         await Promise.all(promises);
@@ -516,14 +514,10 @@ export class PerforceSCMProvider {
         return this.instances.some(inst => inst._model.mayHaveConflictForFile(uri));
     }
 
-    /**
-     * This is the default action when an resource is clicked in the viewlet.
-     * For ADD, AND UNDELETE just show the local file.
-     * For DELETE just show the server file.
-     * For EDIT AND RENAME show the diff window (server on left, local on right).
-     */
-
-    private static async open(resource: Resource, diffType?: DiffType): Promise<void> {
+    private static async open(
+        resource: Resource,
+        diffType?: DiffProvider.DiffType
+    ): Promise<void> {
         if (resource.FileType.base === FileType.BINARY) {
             const uri = Utils.makePerforceDocUri(resource.resourceUri, "fstat", "");
             await workspace
@@ -532,171 +526,6 @@ export class PerforceSCMProvider {
             return;
         }
 
-        if (diffType === undefined) {
-            diffType = resource.isShelved
-                ? DiffType.SHELVE_V_DEPOT
-                : DiffType.WORKSPACE_V_DEPOT;
-        }
-
-        const left = PerforceSCMProvider.getLeftResource(resource, diffType);
-        const right = PerforceSCMProvider.getRightResource(resource, diffType);
-
-        if (!left) {
-            if (!right) {
-                // TODO
-                console.error("Status not supported: " + resource.status.toString());
-                return;
-            }
-            await window.showTextDocument(right);
-            return;
-        }
-        if (!right) {
-            await window.showTextDocument(left.uri);
-            return;
-        }
-        await commands.executeCommand<void>(
-            "vscode.diff",
-            left.uri,
-            right,
-            PerforceSCMProvider.getTitle(resource, left.title, diffType)
-        );
-        return;
-    }
-
-    // Gets the uri for the previous version of the file.
-    private static getLeftResource(
-        resource: Resource,
-        diffType: DiffType
-    ): { title: string; uri: Uri } | undefined {
-        const args = {
-            depot: resource.isShelved,
-            workspace: resource.model.workspaceUri.fsPath
-        };
-
-        if (diffType === DiffType.WORKSPACE_V_SHELVE) {
-            // left hand side is the shelve
-            switch (resource.status) {
-                case Status.ADD:
-                case Status.EDIT:
-                case Status.INTEGRATE:
-                case Status.MOVE_ADD:
-                case Status.BRANCH:
-                    return {
-                        title:
-                            Path.basename(resource.resourceUri.fsPath) +
-                            "@=" +
-                            resource.change,
-                        uri: resource.resourceUri.with({
-                            scheme: "perforce",
-                            query: Utils.makePerforceUriQuery("print", "-q", args),
-                            fragment: "@=" + resource.change
-                        })
-                    };
-                case Status.DELETE:
-                case Status.MOVE_DELETE:
-            }
-        } else {
-            const emptyDoc = Uri.parse("perforce:EMPTY");
-            // left hand side is the depot version
-            switch (resource.status) {
-                case Status.ADD:
-                case Status.BRANCH:
-                    return {
-                        title: Path.basename(resource.resourceUri.fsPath) + "#0",
-                        uri: emptyDoc
-                    };
-                case Status.MOVE_ADD:
-                    // diff against the old file if it is known (always a depot path)
-                    return {
-                        title: resource.fromFile
-                            ? Path.basename(resource.fromFile.fsPath) +
-                              "#" +
-                              resource.fromEndRev
-                            : "Depot Version",
-                        uri: resource.fromFile
-                            ? Utils.makePerforceDocUri(resource.fromFile, "print", "-q", {
-                                  depot: true,
-                                  workspace: resource.model.workspaceUri.fsPath
-                              }).with({ fragment: resource.fromEndRev })
-                            : emptyDoc
-                    };
-                case Status.INTEGRATE:
-                case Status.EDIT:
-                case Status.DELETE:
-                case Status.MOVE_DELETE:
-                    return {
-                        title:
-                            Path.basename(resource.resourceUri.fsPath) +
-                            "#" +
-                            resource.workingRevision,
-                        uri: Utils.makePerforceDocUri(
-                            resource.resourceUri,
-                            "print",
-                            "-q",
-                            args
-                        ).with({ fragment: resource.workingRevision })
-                    };
-            }
-        }
-    }
-
-    // Gets the uri for the current version of the file (or the shelved version depending on the diff type).
-    private static getRightResource(
-        resource: Resource,
-        diffType: DiffType
-    ): Uri | undefined {
-        const emptyDoc = Uri.parse("perforce:EMPTY");
-        if (diffType === DiffType.SHELVE_V_DEPOT) {
-            const args = {
-                depot: resource.isShelved,
-                workspace: resource.model.workspaceUri.fsPath
-            };
-
-            switch (resource.status) {
-                case Status.ADD:
-                case Status.EDIT:
-                case Status.MOVE_ADD:
-                case Status.INTEGRATE:
-                case Status.BRANCH:
-                    return resource.resourceUri.with({
-                        scheme: "perforce",
-                        query: Utils.makePerforceUriQuery("print", "-q", args),
-                        fragment: "@=" + resource.change
-                    });
-            }
-        } else {
-            const exists =
-                !resource.isShelved ||
-                (resource.underlyingUri && fs.existsSync(resource.underlyingUri.fsPath));
-            switch (resource.status) {
-                case Status.ADD:
-                case Status.EDIT:
-                case Status.MOVE_ADD:
-                case Status.INTEGRATE:
-                case Status.BRANCH:
-                    return exists ? resource.underlyingUri ?? emptyDoc : emptyDoc;
-            }
-        }
-    }
-
-    private static getTitle(
-        resource: Resource,
-        leftTitle: string,
-        diffType: DiffType
-    ): string {
-        const basename = Path.basename(resource.resourceUri.fsPath);
-
-        let text = "";
-        switch (diffType) {
-            case DiffType.SHELVE_V_DEPOT:
-                text = leftTitle + " vs " + basename + "@=" + resource.change;
-                break;
-            case DiffType.WORKSPACE_V_SHELVE:
-                text = leftTitle + " vs " + basename + " (workspace)";
-                break;
-            case DiffType.WORKSPACE_V_DEPOT:
-                text = leftTitle + " vs " + basename + " (workspace)";
-        }
-        return text;
+        await DiffProvider.diffDefault(resource, diffType);
     }
 }
